requirements.txt
PySide6>=6.5.0
tinify>=1.6.0
DrissionPage>=4.0.0
Pillow>=10.0.0
